Change directory before running `pip` in setup
If the current working directory in the setup process
happened to contain a `setup.cfg` file (possible if the
user ran `actions/checkout` on a Python package), then
`pip` will install to the `install_scripts` location
specified in said configuration file. This can lead to
situations such as ros-tooling/setup-ros#75, where `colcon`
(and other Python dependencies) get installed in a
location that is not on the `PATH`.

This commit temporarily changes directory to `/` before
updating `pip` and installing Python dependencies, hence
circumventing the installation of Python packages in non-
standard locations.

Signed-off-by: Abrar Rahman Protyasha <abrar@openrobotics.org>
Signed-off-by: aprotyas <aprotyas@u.rochester.edu>